+++ server/routes/news.js
@@ -1,4 +1,4 @@
-import post from '../controller/post'
+import news from '../controller/news.js'
 var express = require('express');
 var router = express.Router();
 var multer = require('multer') // 
@@ -14,8 +14,8 @@ var storage = multer.diskStorage({
 var upload = multer({ storage: storage }); // 3-2
 
 /* GET users listing. */
-router.get('/', post.getAll);
-router.get('/:id', post.getPost);
+router.get('/', news.getNews);
+router.get('/:aId', news.getNewsId);
 // router.post('/', upload.single('images'), post.createPost);
 // router.put('/:id', upload.single('images'), post.updatePost);
 // router.delete('/:id', post.deletePost);
